// convex/_generated/api.d.ts
/* eslint-disable */
/**
 * Generated `api` utility.
 *
 * THIS CODE IS AUTOMATICALLY GENERATED.
 *
 * To regenerate, run `npx convex dev`.
 * @module
 */

import type * as auth from "../auth.js";
import type * as authz from "../authz.js";
import type * as crons from "../crons.js";
import type * as fasts from "../fasts.js";
import type * as http from "../http.js";
import type * as internal_ from "../internal.js";
import type * as plans from "../plans.js";
import type * as push_sender from "../push_sender.js";
import type * as reminders from "../reminders.js";
import type * as settings from "../settings.js";
import type * as users from "../users.js";
import type * as webpush from "../webpush.js";

import type {
  ApiFromModules,
  FilterApi,
  FunctionReference,
} from "convex/server";

declare const fullApi: ApiFromModules<{
  auth: typeof auth;
  authz: typeof authz;
  crons: typeof crons;
  fasts: typeof fasts;
  http: typeof http;
  internal: typeof internal_;
  plans: typeof plans;
  push_sender: typeof push_sender;
  reminders: typeof reminders;
  settings: typeof settings;
  users: typeof users;
  webpush: typeof webpush;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;

/**
 * A utility for referencing Convex functions in your app's internal API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = internal.myModule.myFunction;
 * ```
 */
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;

export declare const components: {};